+ RAW поддержка CatsGuard без сторонней библиотеки. К доработке: админ хендлер с коммандами, illegalException - дропы(дц почему то не работает) * Фикс дюпа через телепорт и стекуемые предметы * Фикс определения цвета ника и титула(очередность в enterworld) + Метод определения перевеса для инстанта L2PcInstance * Фикс бага с дропом предметов на землю через переполненный инвентарь и пета * Локализация Ивентов(не полная)
git-svn-id: https://subversion.assembla.com/svn/scoria4@109 0f1168a5-6793-4626-a6db-a0a741385511

+++ interlude/core/GameServer/java/com/l2scoria/gameserver/model/actor/instance/L2MonsterInstance.java
@@ -228,7 +228,7 @@ public class L2MonsterInstance extends L2Attackable
 			_returnHomeTask.cancel(true);
 		}
 
-		if(Config.ALT_MINION_REMOVE || this.isRaid)
+		if(Config.ALT_MINION_REMOVE && this.isRaid)
 		{
 			deleteSpawnedMinions(false);
 		}
